+++ src/hyp3_isce2/__main__.py
@@ -23,6 +23,7 @@ def main():
         default='insar_tops_burst',
         help='Select the HyP3 entrypoint to use',  # HyP3 entrypoints are specified in `pyproject.toml`
     )
+    parser.add_argument('++omp-num-threads', type=int, help='The number of OpenMP threads to use for parallel regions')
 
     args, unknowns = parser.parse_known_args()
 
@@ -38,6 +39,10 @@ def main():
     # NOTE: Will need to update to `entry_points(group='hyp3', name=args.process)` when updating to python 3.10
     eps = entry_points()['hyp3']
     (process_entry_point,) = {process for process in eps if process.name == args.process}
+
+    if args.omp_num_threads:
+        os.environ['OMP_NUM_THREADS'] = str(args.omp_num_threads)
+
     sys.argv = [args.process, *unknowns]
     sys.exit(process_entry_point.load()())