+++ src/hyp3_isce2/insar_tops_burst.py
@@ -1,22 +1,31 @@
 """Create a single-burst Sentinel-1 geocoded unwrapped interferogram using ISCE2's TOPS processing workflow"""
 
 import argparse
+import json
 import logging
 import os
 import site
+import subprocess
 import sys
+from collections import namedtuple
 from pathlib import Path
-from shutil import make_archive
+from shutil import copyfile, make_archive
 
 from hyp3lib.aws import upload_file_to_s3
 from hyp3lib.get_orb import downloadSentinelOrbitFile
-from hyp3lib.image import create_thumbnail
+from osgeo import gdal
 
 from hyp3_isce2 import topsapp
-from hyp3_isce2.burst import BurstParams, download_bursts, get_isce2_burst_bbox, get_region_of_interest
+from hyp3_isce2.burst import (
+    BurstParams,
+    download_bursts,
+    get_isce2_burst_bbox,
+    get_product_name,
+    get_region_of_interest,
+)
 from hyp3_isce2.dem import download_dem_for_isce2
 from hyp3_isce2.s1_auxcal import download_aux_cal
-
+from hyp3_isce2.utils import make_browse_image, utm_from_lon_lat
 
 log = logging.getLogger(__name__)
 
@@ -90,9 +99,78 @@ def insar_tops_burst(
 
     topsapp.run_topsapp_burst(start='startup', end='preprocess', config_xml=config_path)
     topsapp.swap_burst_vrts()
-    topsapp.run_topsapp_burst(start='computeBaselines', end='geocode', config_xml=config_path)
+    topsapp.run_topsapp_burst(start='computeBaselines', end='unwrap2stage', config_xml=config_path)
+    copyfile('merged/z.rdr.full.xml', 'merged/z.rdr.full.vrt.xml')
+    topsapp.run_topsapp_burst(start='geocode', end='geocode', config_xml=config_path)
 
     return Path('merged')
+
+
+# TODO add more parameters
+# TODO does the format need to be the same as for our INSAR_GAMMA products?
+# TODO unit test
+def make_parameter_file(out_path: Path, reference_scene: str, secondary_scene: str) -> None:
+    output = {
+        'reference_scene': reference_scene,
+        'secondary_scene': secondary_scene,
+    }
+    with out_path.open('w') as f:
+        json.dump(output, f)
+
+
+def translate_outputs(product_dir: Path, product_name: str):
+    """Translate ISCE outputs to a standard GTiff format with a UTM projection
+
+    Args:
+        product_dir: Path to the ISCE merge directory
+        product_name: Name of the product
+    """
+    ISCE2Dataset = namedtuple('ISCE2Dataset', ['name', 'suffix', 'band'])
+    datasets = [
+        ISCE2Dataset('filt_topophase.unw.geo', 'unw_phase', 2),
+        ISCE2Dataset('phsig.cor.geo', 'corr', 1),
+        ISCE2Dataset('z.rdr.full.geo', 'dem', 1),
+        ISCE2Dataset('filt_topophase.unw.conncomp.geo', 'conncomp', 1),
+    ]
+
+    for dataset in datasets:
+        out_file = str(Path(product_name) / f'{product_name}_{dataset.suffix}.tif')
+        in_file = str(product_dir / dataset.name)
+
+        gdal.Translate(
+            destName=out_file,
+            srcDS=in_file,
+            bandList=[dataset.band],
+            format='GTiff',
+            noData=0,
+            creationOptions=['TILED=YES', 'COMPRESS=LZW', 'NUM_THREADS=ALL_CPUS'],
+        )
+
+    wrapped_phase = ISCE2Dataset('filt_topophase.flat.geo', 'wrapped_phase', 1)
+    cmd = (
+        'gdal_calc.py '
+        f'--outfile {product_name}/{product_name}_{wrapped_phase.suffix}.tif '
+        f'-A {product_dir / wrapped_phase.name} '
+        '--calc angle(A) --type Float32 --format GTiff --NoDataValue=0 '
+        '--creation-option TILED=YES --creation-option COMPRESS=LZW --creation-option NUM_THREADS=ALL_CPUS'
+    )
+    subprocess.check_call(cmd.split(' '))
+
+    ds = gdal.Open(str(product_dir / 'filt_topophase.unw.geo'))
+    geotransform = ds.GetGeoTransform()
+    del ds
+
+    epsg = utm_from_lon_lat(geotransform[0], geotransform[3])
+    files = [str(path) for path in Path(product_name).glob('*.tif')]
+    for file in files:
+        gdal.Warp(
+            file,
+            file,
+            dstSRS=f'epsg:{epsg}',
+            creationOptions=['TILED=YES', 'COMPRESS=LZW', 'NUM_THREADS=ALL_CPUS'],
+        )
+
+    make_browse_image(f'{product_name}/{product_name}_unw_phase.tif', f'{product_name}/{product_name}_unw_phase.png')
 
 
 def main():
@@ -128,18 +206,23 @@ def main():
 
     log.info('ISCE2 TopsApp run completed successfully')
 
+    product_name = get_product_name(
+        args.reference_scene,
+        args.secondary_scene,
+        args.reference_burst_number,
+        args.secondary_burst_number,
+        args.swath_number,
+        args.polarization,
+    )
+
+    Path(product_name).mkdir(parents=True, exist_ok=True)
+    translate_outputs(product_dir, product_name)
+    make_parameter_file(
+        Path(f'{product_name}/{product_name}.json'),
+        args.reference_scene,
+        args.secondary_scene,
+    )
+    product_file = make_archive(base_name=product_name, format='zip', base_dir=product_name)
+
     if args.bucket:
-        reference_name = (
-            f'{args.reference_scene}_IW{args.swath_number}_{args.polarization}_{args.reference_burst_number}'
-        )
-        secondary_name = (
-            f'{args.secondary_scene}_IW{args.swath_number}_{args.polarization}_{args.secondary_burst_number}'
-        )
-        base_name = f'{reference_name}x{secondary_name}'
-        product_file = make_archive(base_name=base_name, format='zip', base_dir=product_dir)
-        upload_file_to_s3(product_file, args.bucket, args.bucket_prefix)
-        browse_images = product_file.with_suffix('.png')
-        for browse in browse_images:
-            thumbnail = create_thumbnail(browse)
-            upload_file_to_s3(browse, args.bucket, args.bucket_prefix)
-            upload_file_to_s3(thumbnail, args.bucket, args.bucket_prefix)
+        upload_file_to_s3(Path(product_file), args.bucket, args.bucket_prefix)