+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.3.0]
+### Added
+* `insar_tops` workflow for processing of full Sentinel-1 SLCs.
+* Ability to pass EarthData username and password as environment variables to workflows. This allows the credentials to be passed to the Docker container via the `-e` option.
+* Generation of output geotiff products with sensible names in subfolder that is also added as a zip archive.
+* Generation of a product browse image based on the unwrapped phase geotiff.
+* The `++omp-num-threads` parameter for control of the number of threads used when ISCE2 calls OpenMP functionality.
+
 ## [0.2.1]
 ### Added
 * The `get_isce2_burst_bbox` function to calculate burst bounding boxes using ISCE2 directly to fix inaccurate generation of burst bounding boxes.